+++ owncloudApp/src/main/java/com/owncloud/android/presentation/security/passcode/PassCodeActivity.kt
@@ -9,8 +9,10 @@
  * @author Abel García de Prada
  * @author Juan Carlos Garrote Gascón
  * @author David Crespo Ríos
+ * @author  Aitor Ballesteros Pavón
+ *
  * Copyright (C) 2011 Bartek Przybylski
- * Copyright (C) 2021 ownCloud GmbH.
+ * Copyright (C) 2024 ownCloud GmbH.
  * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -118,6 +120,7 @@ class PassCodeActivity : AppCompatActivity(), NumberKeyboardListener, EnableBiom
                 binding.explanation.visibility = View.INVISIBLE
                 supportActionBar?.setDisplayHomeAsUpEnabled(false) //Don´t show the back arrow
             }
+
             ACTION_CREATE -> { //Create a new password
                 if (confirmingPassCode) {
                     //the app was in the passcode confirmation
@@ -136,15 +139,18 @@ class PassCodeActivity : AppCompatActivity(), NumberKeyboardListener, EnableBiom
                         intent.extras?.getBoolean(EXTRAS_MIGRATION) == true -> {
                             supportActionBar?.setDisplayHomeAsUpEnabled(false)
                         }
+
                         intent.extras?.getBoolean(EXTRAS_LOCK_ENFORCED) == true -> {
                             supportActionBar?.setDisplayHomeAsUpEnabled(false)
                         }
+
                         else -> {
                             supportActionBar?.setDisplayHomeAsUpEnabled(true)
                         }
                     }
                 }
             }
+
             ACTION_REMOVE -> { // Remove password
                 // pass code preference has just been disabled in Preferences;
                 // will confirm user knows pass code, then remove it
@@ -152,6 +158,7 @@ class PassCodeActivity : AppCompatActivity(), NumberKeyboardListener, EnableBiom
                 binding.explanation.visibility = View.INVISIBLE
                 supportActionBar?.setDisplayHomeAsUpEnabled(true)
             }
+
             else -> {
                 throw IllegalArgumentException(R.string.illegal_argument_exception_message.toString() + " ")
             }
@@ -231,12 +238,14 @@ class PassCodeActivity : AppCompatActivity(), NumberKeyboardListener, EnableBiom
                         else -> actionCheckError()
                     }
                 }
+
                 PasscodeAction.REMOVE -> {
                     when (status.type) {
                         PasscodeType.OK -> actionRemoveOk()
                         else -> actionRemoveError()
                     }
                 }
+
                 PasscodeAction.CREATE -> {
                     when (status.type) {
                         PasscodeType.NO_CONFIRM -> actionCreateNoConfirm()
@@ -421,6 +430,7 @@ class PassCodeActivity : AppCompatActivity(), NumberKeyboardListener, EnableBiom
             BiometricStatus.ENABLED_BY_USER -> {
                 passCodeViewModel.setBiometricsState(enabled = true)
             }
+
             BiometricStatus.DISABLED_BY_USER -> {
                 passCodeViewModel.setBiometricsState(enabled = false)
             }
@@ -434,12 +444,49 @@ class PassCodeActivity : AppCompatActivity(), NumberKeyboardListener, EnableBiom
             ACTION_REMOVE -> {
                 return PasscodeAction.REMOVE
             }
+
             ACTION_CREATE -> {
                 return PasscodeAction.CREATE
             }
+
             else -> {
                 return PasscodeAction.CHECK
             }
+        }
+    }
+
+    override fun onKeyUp(keyCode: Int, event: KeyEvent?): Boolean {
+        return when (keyCode) {
+            KeyEvent.KEYCODE_TAB -> {
+                findViewById<View>(R.id.lock_time).requestFocus()
+                true
+            }
+
+            KeyEvent.KEYCODE_DPAD_DOWN -> {
+                if (!findViewById<View>(R.id.numberKeyboard).hasFocus()) {
+                    findViewById<View>(R.id.numberKeyboard).requestFocus()
+                }
+                true
+            }
+
+            in KeyEvent.KEYCODE_0..KeyEvent.KEYCODE_9 -> {
+                val number = keyCode - KeyEvent.KEYCODE_0
+                passCodeViewModel.onNumberClicked(number)
+                true
+            }
+
+            KeyEvent.KEYCODE_DEL -> {
+                passCodeViewModel.onBackspaceClicked()
+                true
+            }
+
+            KeyEvent.KEYCODE_ESCAPE -> {
+                PassCodeManager.onActivityStopped(this)
+                super.onBackPressed()
+                true
+            }
+
+            else -> super.onKeyUp(keyCode, event)
         }
     }